move TreeNodeContainer props to types/interfaces definition file
